add quality to card cache logic and add quality and global id to subcard

+++ rmtg/src/mat.rs
@@ -40,7 +40,7 @@ pub fn create_mats(
             transform,
             right,
             PLAYER[i],
-            Peer::new(i),
+            Peer::new(i as u64),
         );
     }
 }
